# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "Capabilities": {
        "properties": {
          "features": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CapabilityState"
            },
            "description": "Collection features that are not individual HTTP operations",
            "title": "Features",
            "type": "object"
          },
          "operations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OperationCapability"
            },
            "description": "All native operations, keyed by METHOD /path",
            "title": "Operations",
            "type": "object"
          },
          "stats": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Stats",
            "type": "object"
          },
          "versions": {
            "$ref": "#/components/schemas/Versions"
          }
        },
        "required": [
          "versions",
          "operations",
          "features"
        ],
        "title": "Capabilities",
        "type": "object"
      },
      "CapabilityState": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "setting": {
            "description": "Setting controlling this capability, when applicable",
            "title": "Setting",
            "type": "string"
          },
          "status": {
            "default": "available",
            "description": "Effective support after checking the running Anki and settings; request validation still applies",
            "enum": [
              "available",
              "disabled",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "CapabilityState",
        "type": "object"
      },
...
      "CollectionActionResult": {
        "description": "Result of a collection-level command that either worked or raised.",
        "properties": {
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "CollectionActionResult",
        "type": "object"
      },
...
      "CollectionMeta": {
        "description": "Collection-level facts a client can't read from any row: whether FSRS is\nenabled (one collection-wide switch - NOT per deck or per preset, even\nthough Anki's deck-options screen hosts the toggle), and the Anki version\nfor feature detection.",
        "properties": {
          "anki_version": {
            "title": "Anki Version",
            "type": "string"
          },
          "fsrs": {
            "title": "Fsrs",
            "type": "boolean"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "fsrs",
          "anki_version"
        ],
        "title": "CollectionMeta",
        "type": "object"
      },
...
      "ExportRequest": {
        "properties": {
          "deck": {
            "description": "Name of the deck to export",
            "title": "Deck",
            "type": "string"
          },
          "path": {
            "description": "Destination .apkg path, on the Anki machine",
            "title": "Path",
            "type": "string"
          },
          "with_media": {
            "default": true,
            "description": "Bundle referenced media files",
            "title": "With Media",
            "type": "boolean"
          },
          "with_scheduling": {
            "default": false,
            "description": "Include due dates and review history",
            "title": "With Scheduling",
            "type": "boolean"
          }
        },
        "required": [
          "deck",
          "path"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportPreferences": {
        "properties": {
          "options": {
            "$ref": "#/components/schemas/ImportOptions"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "unsupported_options": {
            "items": {
              "type": "string"
            },
            "title": "Unsupported Options",
            "type": "array"
          }
        },
        "required": [
          "options"
        ],
        "title": "ImportPreferences",
        "type": "object"
      },
      "ImportRequest": {
        "additionalProperties": false,
        "description": "Omitted or null options use Anki's saved import preferences.",
        "properties": {
          "merge_notetypes": {
            "description": "Merge compatible note types. Omit or null: use Anki's saved choice.",
            "nullable": true,
            "title": "Merge Notetypes",
            "type": "boolean"
          },
          "path": {
            "description": "Source .apkg path, on the Anki machine",
            "title": "Path",
            "type": "string"
          },
          "update_notes": {
            "description": "When to update existing notes. Omit or null: use Anki's saved choice.",
            "enum": [
              "if_newer",
              "always",
              "never"
            ],
            "nullable": true,
            "title": "Update Notes",
            "type": "string"
          },
          "update_notetypes": {
            "description": "When to update existing note types. Omit or null: use Anki's saved choice.",
            "enum": [
              "if_newer",
              "always",
              "never"
            ],
            "nullable": true,
            "title": "Update Notetypes",
            "type": "string"
          },
          "with_deck_configs": {
            "description": "Import deck presets. Omit or null: use Anki's saved choice. Explicit values require backend support; inspect /v1/collection/import-options.",
            "nullable": true,
            "title": "With Deck Configs",
            "type": "boolean"
          },
          "with_scheduling": {
            "description": "Import due dates and review history. Omit or null: use Anki's saved choice.",
            "nullable": true,
            "title": "With Scheduling",
            "type": "boolean"
          }
        },
        "required": [
          "path"
        ],
        "title": "ImportRequest",
        "type": "object"
      },
      "ImportResult": {
        "properties": {
          "imported": {
            "default": 0,
            "title": "Imported",
            "type": "integer"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "updated": {
            "default": 0,
            "title": "Updated",
            "type": "integer"
          }
        },
        "title": "ImportResult",
        "type": "object"
      },
...
      "JobInfo": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "result": {
            "title": "Result",
            "type": "object"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "stats"
        ],
        "title": "JobInfo",
        "type": "object"
      },
...
      "JobSubmitted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "stats"
        ],
        "title": "JobSubmitted",
        "type": "object"
      },
...
      "ProfileList": {
        "description": "Every profile Anki knows about, and which one is open.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "string"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProfileList",
        "type": "object"
      },
      "ProfileLoad": {
        "properties": {
          "name": {
            "description": "Profile to switch to",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProfileLoad",
        "type": "object"
      },
      "ProfileLoadResult": {
        "properties": {
          "loaded": {
            "title": "Loaded",
            "type": "boolean"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "loaded"
        ],
        "title": "ProfileLoadResult",
        "type": "object"
      },
...
      "SyncResult": {
        "properties": {
          "server_message": {
            "default": "",
            "title": "Server Message",
            "type": "string"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "status"
        ],
        "title": "SyncResult",
        "type": "object"
      },
...
      "Versions": {
        "properties": {
          "addon": {
            "default": "0.2.0",
            "description": "Tsunagi release version",
            "title": "Addon",
            "type": "string"
          },
          "anki": {
            "description": "Running Anki version",
            "title": "Anki",
            "type": "string"
          },
          "api": {
            "default": "v1",
            "description": "Native API contract identifier",
            "title": "Api",
            "type": "string"
          }
        },
        "required": [
          "anki"
        ],
        "title": "Versions",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Required when an API key is configured in Tsunagi settings.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Alternative to the X-API-Key header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT"
    },
    "title": "Tsunagi"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "description": "One report for all native operations and collection features. Each has an effective status: available, disabled in settings, or unsupported by this Anki. Conditional options carry their own status, reason and setting. Availability does not bypass input validation, authentication or transient collection/GUI state. Requires an open collection; /v1/health remains a lightweight liveness check. AnkiConnect actions are listed separately at /actions.",
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Native API capabilities",
        "tags": [
          "Collection"
        ],
        "x-permission": "read:collection"
      }
    },
...
    "/v1/collection": {
      "get": {
        "description": "Collection-wide facts with no row of their own: whether FSRS is enabled (one switch for the whole collection - per-deck and per-preset `desired_retention` values exist either way, the scheduler just ignores them while this is off), and the running Anki version for feature detection.",
        "operationId": "getCollectionMeta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionMeta"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Collection metadata",
        "tags": [
          "Collection"
        ],
        "x-permission": "read:collection"
      }
    },
    "/v1/collection/import-options": {
      "get": {
        "description": "Returns Anki's saved import choices for the open collection. Unsupported options are listed separately and have null values. Reading does not import a file or change preferences. Pass explicit options to /v1/collection:import to keep a reviewed choice fixed even if preferences change before the import.",
        "operationId": "getImportOptions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreferences"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Read saved package import options",
        "tags": [
          "Collection"
        ],
        "x-permission": "read:collection"
      }
    },
    "/v1/collection:check-database": {
      "post": {
        "description": "Runs Anki's Check Database, which rebuilds indexes and fixes inconsistencies.",
        "operationId": "checkDatabase",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionActionResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Check the database",
        "tags": [
          "Collection"
        ],
        "x-permission": "manage"
      }
    },
    "/v1/collection:export": {
      "post": {
        "description": "Writes one deck to an .apkg file. The path is resolved on the machine running Anki, not the caller's.",
        "operationId": "exportPackage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Export a deck",
        "tags": [
          "Collection"
        ],
        "x-permission": "manage"
      }
    },
    "/v1/collection:import": {
      "post": {
        "description": "Merges an .apkg into the current collection. The path is resolved on the machine running Anki. Imports immediately without an options dialog. Omitted or null options use Anki's saved import preferences, available at GET /v1/collection/import-options. Explicit values override those choices. Anki remembers the resulting options after a successful import. Unsupported explicit options are rejected before importing. Returns the result with HTTP 200 if the import completes within the operation timeout; otherwise returns HTTP 202 with a job_id and Location header. Poll GET /v1/jobs/{job_id}; the same import continues without restarting. Import jobs cannot be aborted through the API. Only one import or FSRS job may be active at once (409 otherwise). Jobs are kept in memory only.",
        "operationId": "importPackage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmitted"
                }
              }
            },
            "description": "Import still running; poll the job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Import a package",
        "tags": [
          "Collection"
        ],
        "x-permission": "manage"
      }
    },
    "/v1/collection:reload": {
      "post": {
        "deprecated": true,
        "description": "This endpoint performs no reload. It remains available for existing clients and returns success when a collection is open. After a completed collection operation, query the data directly; no reload step is required. This request does not clear caches or reopen the collection.",
        "operationId": "reloadCollection",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionActionResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reload the collection (deprecated no-op)",
        "tags": [
          "Collection"
        ],
        "x-permission": "manage"
      }
    },
    "/v1/collection:sync": {
      "post": {
        "description": "Runs a sync against AnkiWeb using the credentials already configured in Anki. Errors if no sync account is set up, or if the collection needs a full upload or download - that requires Anki's own UI.",
        "operationId": "syncCollection",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Sync the collection",
        "tags": [
          "Collection"
        ],
        "x-permission": "sync"
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "Status, best-effort progress while running, and the result or error once terminal. Finished jobs are kept in memory until evicted, not persisted. Import jobs report their result or error but do not expose progress or support abort.",
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Poll a job",
        "tags": [
          "Jobs"
        ],
        "x-permission": "read:collection"
      }
    },
...
    "/v1/profiles": {
      "get": {
        "description": "Every Anki profile, and which one is currently open.",
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "List profiles",
        "tags": [
          "Collection"
        ],
        "x-permission": "read:collection"
      }
    },
    "/v1/profiles:load": {
      "post": {
        "description": "Schedules closing the current profile and opening another. `loaded` confirms acceptance, not completion, and is false for an unknown name. The server restarts during the switch; requests may encounter 503 or a connection interruption. Reconnect and GET /v1/profiles to confirm the active profile.",
        "operationId": "loadProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileLoad"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileLoadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Switch profile",
        "tags": [
          "Collection"
        ],
        "x-permission": "manage"
      }
    },
...
  },
...
}